src/modules/indel_primer/requirements.txt
logzero==1.3.1
python-dotenv==0.19.1
sentry-sdk==1.7.2
itsdangerous==2.0.1
Jinja2==3.0.3
